Collection of Labs
These are all my lab attempts during the Fall semester of 2014. These
files also include work that is not my own but was necessary for the
programs to operate - like jQuery. Lab 5, in particular, was an original
program that I altered in order to get a handle on D3.  All alterations
should be noted in the comments contained in histogram.js.